requirement.txt
alabaster==0.7.12
appdirs==1.4.4
APScheduler==3.6.3
argh==0.26.2
astroid @ file:///C:/ci/astroid_1592481955828/work
async-generator==1.10
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1596578164842/work
Babel @ file:///tmp/build/80754af9/babel_1607110387436/work
backcall==0.2.0
bcrypt @ file:///C:/ci/bcrypt_1597918112552/work
black==19.10b0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255207413/work
chardet @ file:///C:/ci/chardet_1607706910910/work
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1607639129468/work
decorator==4.4.2
defusedxml==0.6.0
dialogflow==1.1.0
diff-match-patch @ file:///tmp/build/80754af9/diff-match-patch_1594828741838/work
docutils==0.16
entrypoints==0.3
feedparser==6.0.2
flake8 @ file:///tmp/build/80754af9/flake8_1601911421857/work
Flask==1.1.2
future==0.18.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.24.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imagesize==1.2.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
intervaltree @ file:///tmp/build/80754af9/intervaltree_1598376443606/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1604101327119/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1606932641827/work
Jinja2==2.11.2
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///C:/ci/jupyter_core_1606130908768/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
keyring @ file:///C:/ci/keyring_1609335582560/work
lazy-object-proxy @ file:///C:/b/work
logging==0.4.9.6
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
mypy-extensions==0.4.3
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
numpydoc @ file:///tmp/build/80754af9/numpydoc_1605117425582/work
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
paramiko @ file:///tmp/build/80754af9/paramiko_1598886428689/work
parso==0.7.0
pathspec==0.7.0
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pluggy==0.13.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.14.0
psutil @ file:///C:/ci/psutil_1598352273540/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1598885001695/work
pyflakes==2.2.0
Pygments @ file:///tmp/build/80754af9/pygments_1607368905949/work
pylint @ file:///C:/ci/pylint_1598605958085/work
pyls-black @ file:///tmp/build/80754af9/pyls-black_1607553132291/work
pyls-spyder @ file:///tmp/build/80754af9/pyls-spyder_1608134179673/work
PyNaCl @ file:///C:/ci/pynacl_1595009241355/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1600278539111/work
python-language-server @ file:///tmp/build/80754af9/python-language-server_1607972495879/work
python-telegram-bot==13.1
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
pywin32==227
pywin32-ctypes @ file:///C:/ci/pywin32-ctypes_1594392691209/work
PyYAML==5.3.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtAwesome @ file:///tmp/build/80754af9/qtawesome_1602272867890/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
regex @ file:///C:/ci/regex_1606691216163/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
rope @ file:///tmp/build/80754af9/rope_1602264064449/work
rsa==4.6
Rtree==0.9.4
sgmllib3k==1.0.0
six @ file:///C:/ci/six_1605205426665/work
snowballstemmer==2.0.0
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1606865132123/work
Sphinx @ file:///tmp/build/80754af9/sphinx_1608836605859/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder @ file:///C:/b/work
spyder-kernels @ file:///C:/ci/spyder-kernels_1608560839696/work
testpath==0.4.4
three-merge @ file:///tmp/build/80754af9/three-merge_1607553261110/work
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
tzlocal==2.1
ujson @ file:///C:/ci/ujson_1602523369547/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
watchdog @ file:///C:/ci/watchdog_1606939375616/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.11.2
yapf @ file:///tmp/build/80754af9/yapf_1593528177422/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
